Runtime/RTAcpi*: Parser updates, the code is now able to parse vbox.dsl successfully and generate an AST, bugref:10733 [build fix]
svn:sync-xref-src-repo-rev: r167503

+++ src/VBox/Runtime/common/acpi/acpi.cpp
@@ -1,4 +1,4 @@
-/* $Id: acpi.cpp 108349 2025-02-12 15:30:37Z alexander.eichner@oracle.com $ */
+/* $Id: acpi.cpp 108351 2025-02-12 15:38:05Z alexander.eichner@oracle.com $ */
 /** @file
  * IPRT - Advanced Configuration and Power Interface (ACPI) Table generation API.
  */
@@ -1740,7 +1740,7 @@ static int rtAcpiResourceAddQWordAddressRange(PRTACPIRESINT pThis, uint8_t bType
                                               uint64_t u64Granularity, uint64_t u64Length,
                                               const char *pszRsrcSrc, uint8_t bRsrcIndex)
 {
-    size_t cchRsrcSrc = pszRsrcSrc ? strlen(pszRsrcSrc) + 2 : 0;
+    uint32_t cchRsrcSrc = pszRsrcSrc ? (uint32_t)strlen(pszRsrcSrc) + 2 : 0;
     uint8_t *pb = rtAcpiResBufEnsureSpace(pThis, 3 + 43 + cchRsrcSrc);
     if (!pb)
         return VERR_NO_MEMORY;
